[INJIWEB-1379] handle the edge cases noticed in login flow
Description: If the session is valid but the displayName is removed from localStorage manually added the logic to fetch the details from backend and also if the session is removed from the redis manually and reload the app we shouldn't show the user name in the UI as it the name is still available in the local storage. Basically we should rely on the session status but not on whether the user details available in storage or not.

Signed-off-by: PuBHARGAVI <46226958+PuBHARGAVI@users.noreply.github.com>