[INJIWEB-1557] keeping limitations in mind making changes in the code to store the downloadInProgress and current session details in the app context itself
Description: If the user closes the tab or reload the tab then the component will not be able to listen to the response of the already initiated download call and UI cannot update the state and will stuck in loading state which is not a valid scenario. To handle these scenraios we have to send an event from the backend to the frontend when download is completed and based on that we have to update the UI. We are going to handle this in future so for now we are storing details in app context itself so the success or error toaster messages for the previous download calls will be shown only in the same tab until the user reloads the tab.

Signed-off-by: PuBHARGAVI <46226958+PuBHARGAVI@users.noreply.github.com>